requirements.txt
neo4j
qdrant-client
langchain
langchain-community
langchain-core
langchain-openai
langchain-experimental
sentence-transformers
pydantic
python-dotenv
django>=5.0
djangorestframework
django-cors-headers
